+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.0.6]
 ### Added 
 - GitHub actions for publishing builds to GitHub and CurseForge
+- Support for [Ex Nihilo Sequentia] (Crucible, Fluid Item Transformation, Fluid On Top, Hammer, Sieve)
 
 ### Removed
 - CurseGradle, it just suddenly broke.
@@ -421,7 +424,8 @@ Added support for 18 more mods
 - Added support for Extended Crafting
     - Combintation Core, Compressor, Ender Crafter and Tiered Tables
 
-[Unreleased]: https://github.com/TagnumElite/ProjectE-Integration/compare/4.0.5...HEAD
+[Unreleased]: https://github.com/TagnumElite/ProjectE-Integration/compare/4.0.6...HEAD
+[4.0.6]: https://github.com/TagnumElite/ProjectE-Integration/compare/4.0.5...4.0.6
 [4.0.5]: https://github.com/TagnumElite/ProjectE-Integration/compare/4.0.4...4.0.5
 [4.0.4]: https://github.com/TagnumElite/ProjectE-Integration/compare/4.0.3...4.0.4
 [4.0.3]: https://github.com/TagnumElite/ProjectE-Integration/compare/4.0.2...4.0.3
@@ -463,6 +467,7 @@ Added support for 18 more mods
 [Applied Energistics]: https://www.curseforge.com/minecraft/mc-mods/applied-energistics-2
 [Compact Crafting]: https://www.curseforge.com/minecraft/mc-mods/compact-crafting
 [Create]: https://www.curseforge.com/minecraft/mc-mods/create
+[Ex Nihilo Sequentia]: https://www.curseforge.com/minecraft/mc-mods/ex-nihilo-sequentia
 [Extended Crafting]: https://www.curseforge.com/minecraft/mc-mods/extended-crafting
 [Immersive Engineering]: https://www.curseforge.com/minecraft/mc-mods/immersive-engineering
 [Woot]: https://www.curseforge.com/minecraft/mc-mods/woot